Added the plot function to the header
