+++ code/project_cpp.d
@@ -1,6 +1,40 @@
 
 # DO NOT DELETE
 
-./project_cpp.so: tempTrender.h
+./project_cpp.so: tempTrender.h /usr/include/root/TF1.h
+./project_cpp.so: /usr/include/root/TFormula.h /usr/include/root/TNamed.h
+./project_cpp.so: /usr/include/root/TObject.h /usr/include/root/Rtypes.h
+./project_cpp.so: /usr/include/root/RConfig.h /usr/include/root/RVersion.h
+./project_cpp.so: /usr/include/root/DllImport.h /usr/include/root/Rtypeinfo.h
+./project_cpp.so: /usr/include/root/snprintf.h /usr/include/root/strlcpy.h
+./project_cpp.so: /usr/include/root/TGenericClassInfo.h
+./project_cpp.so: /usr/include/root/TSchemaHelper.h
+./project_cpp.so: /usr/include/root/TStorage.h
+./project_cpp.so: /usr/include/root/TVersionCheck.h
+./project_cpp.so: /usr/include/root/Riosfwd.h /usr/include/root/TBuffer.h
+./project_cpp.so: /usr/include/root/TString.h /usr/include/root/TMathBase.h
+./project_cpp.so: /usr/include/root/TBits.h /usr/include/root/TObjArray.h
+./project_cpp.so: /usr/include/root/TSeqCollection.h
+./project_cpp.so: /usr/include/root/TCollection.h
+./project_cpp.so: /usr/include/root/TIterator.h /usr/include/root/TAttLine.h
+./project_cpp.so: /usr/include/root/TAttFill.h /usr/include/root/TAttMarker.h
+./project_cpp.so: /usr/include/root/Math/ParamFunctor.h
+./project_cpp.so: /usr/include/root/TH1.h /usr/include/root/TAxis.h
+./project_cpp.so: /usr/include/root/TAttAxis.h /usr/include/root/TArrayD.h
+./project_cpp.so: /usr/include/root/TArray.h /usr/include/root/TArrayC.h
+./project_cpp.so: /usr/include/root/TArrayS.h /usr/include/root/TArrayI.h
+./project_cpp.so: /usr/include/root/TArrayF.h /usr/include/root/Foption.h
+./project_cpp.so: /usr/include/root/TVectorFfwd.h
+./project_cpp.so: /usr/include/root/TVectorDfwd.h
+./project_cpp.so: /usr/include/root/TFitResultPtr.h
+./project_cpp.so: /usr/include/root/TStyle.h /usr/include/root/TAttText.h
+./project_cpp.so: /usr/include/root/TMath.h /usr/include/root/TError.h
+./project_cpp.so: /usr/include/root/TCanvas.h /usr/include/root/TPad.h
+./project_cpp.so: /usr/include/root/TVirtualPad.h /usr/include/root/TAttPad.h
+./project_cpp.so: /usr/include/root/TVirtualX.h /usr/include/root/GuiTypes.h
+./project_cpp.so: /usr/include/root/Buttons.h /usr/include/root/TQObject.h
+./project_cpp.so: /usr/include/root/TAttBBox2D.h /usr/include/root/TPoint.h
+./project_cpp.so: /usr/include/root/TAttCanvas.h
+./project_cpp.so: /usr/include/root/TCanvasImp.h /usr/include/root/TGraph.h
 ./project_cpp.so: /usr/include/root/cintdictversion.h /usr/include/root/RVersion.h
 project_cpp__ROOTBUILDVERSION= 5.34/30
